# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * YCIS Monopoly 2026 Game Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Team {
  id: number;
  name: string;
  cash: number;
  /** Board space index (0-39) */
  position: number;
  /** Hex color code for team token */
  color: string;
  /** Team emoji avatar */
  emoji: string;
  /** Cash + property values + set bonuses */
  netWorth: number;
  propertyCount: number;
  createdAt?: string;
}

export interface TeamInput {
  /** @minLength 1 */
  name: string;
  cash?: number;
  position?: number;
  color: string;
  emoji: string;
}

export interface TeamUpdate {
  name?: string;
  cash?: number;
  position?: number;
  color?: string;
  emoji?: string;
}

export type BoardSpaceType = typeof BoardSpaceType[keyof typeof BoardSpaceType];


export const BoardSpaceType = {
  property: 'property',
  chance: 'chance',
  community_chest: 'community_chest',
  tax: 'tax',
  corner: 'corner',
  dice_station: 'dice_station',
} as const;

export interface BoardSpace {
  id: number;
  name: string;
  type: BoardSpaceType;
  position: number;
  /**
     * Monopoly color group (e.g. dark_blue, green)
     * @nullable
     */
  colorGroup: string | null;
  rentValue: number;
  propertyValue: number;
  hasHotel?: boolean;
  /**
     * Team ID of the owner, null if unowned
     * @nullable
     */
  ownerId?: number | null;
  /** @nullable */
  ownerName?: string | null;
  /** @nullable */
  ownerColor?: string | null;
  /** @nullable */
  ownerEmoji?: string | null;
}

export interface OwnershipUpdate {
  /** @nullable */
  ownerId?: number | null;
  hasHotel?: boolean;
}

export type GameEventType = typeof GameEventType[keyof typeof GameEventType];


export const GameEventType = {
  property_taken: 'property_taken',
  rent_paid: 'rent_paid',
  takeover: 'takeover',
  chance: 'chance',
  community_chest: 'community_chest',
  hotel_built: 'hotel_built',
  cash_change: 'cash_change',
  system: 'system',
} as const;

export interface GameEvent {
  id: number;
  message: string;
  type: GameEventType;
  /** @nullable */
  teamId?: number | null;
  /** @nullable */
  teamName?: string | null;
  /** @nullable */
  teamEmoji?: string | null;
  /** @nullable */
  amount?: number | null;
  createdAt: string;
}

export type GameEventInputType = typeof GameEventInputType[keyof typeof GameEventInputType];


export const GameEventInputType = {
  property_taken: 'property_taken',
  rent_paid: 'rent_paid',
  takeover: 'takeover',
  chance: 'chance',
  community_chest: 'community_chest',
  hotel_built: 'hotel_built',
  cash_change: 'cash_change',
  system: 'system',
} as const;

export interface GameEventInput {
  /** @minLength 1 */
  message: string;
  type: GameEventInputType;
  /** @nullable */
  teamId?: number | null;
  /** @nullable */
  amount?: number | null;
}

export type GameStateStatus = typeof GameStateStatus[keyof typeof GameStateStatus];


export const GameStateStatus = {
  lobby: 'lobby',
  active: 'active',
  finished: 'finished',
} as const;

export interface GameState {
  teams: Team[];
  board: BoardSpace[];
  recentEvents: GameEvent[];
  round: number;
  status: GameStateStatus;
}

export interface LeaderboardEntry {
  rank: number;
  teamId: number;
  teamName: string;
  teamEmoji: string;
  teamColor: string;
  cash: number;
  propertyValue: number;
  setBonus: number;
  netWorth: number;
  propertyCount: number;
}

export interface ResetResult {
  success: boolean;
  message: string;
}

export type ListEventsParams = {
limit?: number;
};