# lib/api-zod/src/generated/types/boardSpace.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * YCIS Monopoly 2026 Game Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { BoardSpaceType } from './boardSpaceType';

export interface BoardSpace {
  id: number;
  name: string;
  type: BoardSpaceType;
  position: number;
  /**
     * Monopoly color group (e.g. dark_blue, green)
     * @nullable
     */
  colorGroup: string | null;
  rentValue: number;
  propertyValue: number;
  hasHotel?: boolean;
  /**
     * Team ID of the owner, null if unowned
     * @nullable
     */
  ownerId?: number | null;
  /** @nullable */
  ownerName?: string | null;
  /** @nullable */
  ownerColor?: string | null;
  /** @nullable */
  ownerEmoji?: string | null;
}