# lib/api-zod/src/generated/types/gameEvent.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * YCIS Monopoly 2026 Game Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { GameEventType } from './gameEventType';

export interface GameEvent {
  id: number;
  message: string;
  type: GameEventType;
  /** @nullable */
  teamId?: number | null;
  /** @nullable */
  teamName?: string | null;
  /** @nullable */
  teamEmoji?: string | null;
  /** @nullable */
  amount?: number | null;
  createdAt: string;
}