# lib/api-zod/src/generated/types/gameEventInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * YCIS Monopoly 2026 Game Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { GameEventInputType } from './gameEventInputType';

export interface GameEventInput {
  /** @minLength 1 */
  message: string;
  type: GameEventInputType;
  /** @nullable */
  teamId?: number | null;
  /** @nullable */
  amount?: number | null;
}