# lib/api-zod/src/generated/types/gameState.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * YCIS Monopoly 2026 Game Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { BoardSpace } from './boardSpace';
import type { GameEvent } from './gameEvent';
import type { GameStateStatus } from './gameStateStatus';
import type { Team } from './team';

export interface GameState {
  teams: Team[];
  board: BoardSpace[];
  recentEvents: GameEvent[];
  round: number;
  status: GameStateStatus;
}