# lib/api-zod/src/generated/types/resetResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * YCIS Monopoly 2026 Game Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface ResetResult {
  success: boolean;
  message: string;
}