# lib/api-zod/src/generated/types/team.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * YCIS Monopoly 2026 Game Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface Team {
  id: number;
  name: string;
  cash: number;
  /** Board space index (0-39) */
  position: number;
  /** Hex color code for team token */
  color: string;
  /** Team emoji avatar */
  emoji: string;
  /** Cash + property values + set bonuses */
  netWorth: number;
  propertyCount: number;
  createdAt?: string;
}